Move some functionality out of commands
