+++ src/storage/txn/mod.rs
@@ -21,7 +21,7 @@ use crate::storage::{
     Error as StorageError, Result as StorageResult,
 };
 
-pub use self::commands::{Command, PointGetCommand};
+pub use self::commands::Command;
 pub use self::process::RESOLVE_LOCK_BATCH_SIZE;
 pub use self::scheduler::{Msg, Scheduler};
 pub use self::store::{EntryBatch, TxnEntry, TxnEntryScanner, TxnEntryStore};
